+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### [0.1.10] - 2026-08-17
+- Add filters to statistics view.
+- Added referral codes to users and added visualization to admin view.
+- Add protection to structural states.
+
 ### [0.1.9] - 2026-08-12
 - Adjusted code generation in order creation to include a prefix based on the client type.
 - Optimized order queries to include pagination and caching.